Fix some signed vs unsigned comparison warnings
They only showed up with WORD_SIZE == 8

+++ src/load_object.c
@@ -69,7 +69,7 @@ int load_object(state *S, int fd, UWORD address)
     if (ptr == NULL)
         return -1;
 
-    if (read(fd, ptr, length) != length)
+    if (read(fd, ptr, length) != (ssize_t)length)
         return -3;
 
     return 0;
